Support Symfony 4
This adds support for symfony 4. No actual code changes were necessary
to accommodate this additional support.

The ci pipeline is updated to test both symfony 3 & 4. Deprecation
warnings need to be suppressed for Symfony 4 because we are unable to
fix most of these without dropping support for Symfony 3. Once we drop
symfony 3 support, these deprecations should be resolved.